+++ src/commands/player/index.ts
@@ -1,0 +1,13 @@
+import { AddPlayerCommand } from "./AddPlayer";
+import { DeletePlayerCommand } from "./DeletePlayer";
+import { MovePlayerCommand } from "./MovePlayer";
+import { SetCaptainCommand } from "./SetCaptain";
+import { SetRecentPlayCommand } from "./SetRecentPlay";
+
+export {
+  AddPlayerCommand,
+  DeletePlayerCommand,
+  MovePlayerCommand,
+  SetCaptainCommand,
+  SetRecentPlayCommand,
+};
